+++ starboard/nplb/posix_compliance/posix_string_format_test.cc
@@ -36,7 +36,7 @@ TEST(PosixFormatTest, SunnyDay) {
   int result = Format(destination, SB_ARRAY_SIZE(destination), "a%db%dc%d%s", 1,
                       2, 3, "test");
   size_t expected_length = strlen(kExpected);
-  EXPECT_EQ(expected_length, result);
+  EXPECT_EQ(expected_length, static_cast<size_t>(result));
   for (size_t i = 0; i <= expected_length; ++i) {
     EXPECT_EQ(kExpected[i], destination[i]);
   }
